feat(rich-text-web): cleanup, add hook
